+++ examples/unified/rl/verl/main.py
@@ -10,6 +10,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
+# This package includes code from [https://github.com/volcengine/verl]
+# licensed under the Apache License 2.0. See [https://github.com/volcengine/verl]
+# for details.
 
 from pprint import pprint
 
@@ -46,6 +50,7 @@ def main(config):
         .config(config)
     )
     # 3. Define roles and their resource requirements
+    # train() for workloads with Rendezvous, while run() for other workloads.
     builder.role("actor_rollout").train("workers.ActorWorker").total(gpus)
     builder.role("critic").train("workers.CriticWorker").total(gpus)
     if config.reward_model.enable:
